extracted some common logic into a helper file and seeded test data
