+++ frontend/.env
@@ -1,9 +1,9 @@
 REACT_APP_TITLE=SimpleReport
-REACT_APP_DESCRIPTION=SimpleReport Test Management Application
+REACT_APP_DESCRIPTION="SimpleReport Test Management Application"
 REACT_APP_ICON=usds-32x32.png
 REACT_APP_BACKEND_URL=https://simplereport.gov/
 REACT_APP_BASE_URL=https://simplereport.gov/
 REACT_APP_IS_TRAINING_SITE=false
 INLINE_RUNTIME_CHUNK=false
-REACT_APP_CONTENT_SECURITY_POLICY_DEFAULT_SRC=default-src 'self';
-REACT_APP_CONTENT_SECURITY_POLICY_SCRIPT_SRC=script-src 'self' dap.digitalgov.gov www.google-analytics.com;
+REACT_APP_CONTENT_SECURITY_POLICY_DEFAULT_SRC="default-src 'self';"
+REACT_APP_CONTENT_SECURITY_POLICY_SCRIPT_SRC="script-src 'self' dap.digitalgov.gov www.google-analytics.com;"
