+++ frontend/src/app/supportAdmin/PendingOrganizations/modals/EditOrgModal.tsx
@@ -2,29 +2,18 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React, { useState } from "react";
 import Modal from "react-modal";
 
-import { PendingOrganization } from "../../../generated/graphql";
-import Button from "../../commonComponents/Button/Button";
-import Input from "../../commonComponents/Input";
-import { isFieldValid, isFormValid } from "../../utils/yupHelpers";
-
+import { PendingOrganization } from "../../../../generated/graphql";
+import Button from "../../../commonComponents/Button/Button";
+import Input from "../../../commonComponents/Input";
+import { isFormValid, isFieldValid } from "../../../utils/yupHelpers";
 import {
   PendingOrganizationFormValues,
   pendingOrganizationSchema,
-} from "./utils";
+} from "../utils";
 
-interface Props {
-  organization: PendingOrganization;
-  onClose: () => void;
-  onSubmit: (organization: PendingOrganizationFormValues) => void;
-  isUpdating: boolean;
-}
+import { ModalProps, PendingOrganizationErrors } from "./help";
 
-type PendingOrganizationErrors = Record<
-  keyof PendingOrganizationFormValues,
-  string
->;
-
-const ConfirmOrgVerificationModal: React.FC<Props> = ({
+const EditOrgModal: React.FC<ModalProps> = ({
   organization,
   onClose,
   onSubmit,
@@ -45,6 +34,9 @@ const ConfirmOrgVerificationModal: React.FC<Props> = ({
     adminPhone: "",
   });
 
+  const getValidationStatus = (field: keyof PendingOrganizationFormValues) =>
+    errors[field] ? "error" : undefined;
+
   const validateField = async (field: keyof PendingOrganizationFormValues) => {
     setErrors(
       await isFieldValid({
@@ -55,9 +47,6 @@ const ConfirmOrgVerificationModal: React.FC<Props> = ({
       })
     );
   };
-
-  const getValidationStatus = (field: keyof PendingOrganizationFormValues) =>
-    errors[field] ? "error" : undefined;
 
   const onChange = (field: keyof PendingOrganization) => (
     value: PendingOrganization[typeof field]
@@ -81,7 +70,7 @@ const ConfirmOrgVerificationModal: React.FC<Props> = ({
     formObject: org,
     onChange,
     required: true,
-    disabled: true,
+    disabled: isUpdating,
     validate: validateField,
     errors,
     getValidationStatus,
@@ -104,7 +93,7 @@ const ConfirmOrgVerificationModal: React.FC<Props> = ({
       <div className="border-0 card-container">
         <div className="display-flex flex-justify">
           <h1 className="font-heading-lg margin-top-05 margin-bottom-0">
-            Confirm the following details are ready for verification
+            Edit organization
           </h1>
           <button onClick={onClose} className="close-button" aria-label="Close">
             <span className="fa-layers">
@@ -158,4 +147,4 @@ const ConfirmOrgVerificationModal: React.FC<Props> = ({
   );
 };
 
-export default ConfirmOrgVerificationModal;
+export default EditOrgModal;
