+++ frontend/src/app/supportAdmin/PendingOrganizations/modals/help.tsx
@@ -1,0 +1,14 @@
+import { PendingOrganizationFormValues } from "../utils";
+import { PendingOrganization } from "../../../../generated/graphql";
+
+export interface ModalProps {
+  organization: PendingOrganization;
+  onClose: () => void;
+  onSubmit: (organization: PendingOrganizationFormValues) => void;
+  isUpdating: boolean;
+}
+
+export type PendingOrganizationErrors = Record<
+  keyof PendingOrganizationFormValues,
+  string
+>;
